% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{association}
\alias{association}
\title{GWAS result.}
\format{:
\describe{
  \item{Marker}{molecular marker name}
  \item{Locus}{the chromosome of the marker}
  \item{Site}{the position of the marker}
  \item{p}{p value of the marker.}
}}
\usage{
association
}
\description{
A dataset containing the association file. The variables are as follows:
}
\keyword{datasets}
