% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtf}
\alias{gtf}
\title{Genome annotation file.}
\format{:
\describe{
  \item{the first column}{The name of the sequence. Commonly, this is the chromosome ID or contig ID}
  \item{the second column}{The source column should be a unique label indicating whether the annotation is protein-coding or something else.}
  \item{the third column}{The following feature types are required: 'CDS', 'start_codon', 'stop_codon'. The features '5UTR', '3UTR', 'inter', 'inter_CNS', 'intron_CNS' and 'exon' are optional.}
  \item{the fourth column}{Integer start coordinates of the feature relative to the beginning of the sequence named}
  \item{the fifth co0umn}{Integer end coordinates of the feature relative to the beginning of the sequence named}
  \item{the sixth column}{The score field indicates a degree of confidence in the feature's existence and coordinates.}
  \item{the seventh column}{The strand indicates the annotation is located on forward or reverse strand.}
  \item{the eigth column}{frame}
  \item{the ninth column}{attributes}
}}
\usage{
gtf
}
\description{
A dataset containing the annotation file WITHOUT the column name. The example file is Maize genome 
annotation, and the version is AGPv2.IMPORTANT, the annotation file could be read with header=FALSE.
The variables are as follows:
}
\keyword{datasets}
