% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hapmap2}
\alias{hapmap2}
\title{Genotype file.}
\format{An object of class \code{data.frame} with 2367 rows and 115 columns.}
\usage{
hapmap2
}
\description{
A dataset containing the genotype file. The variables see hapmap_am368
}
\keyword{datasets}
