% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntGenicPlot.R
\name{IntGenicPlot}
\alias{IntGenicPlot}
\title{plot association with LD and annotation at a given gene}
\usage{
IntGenicPlot(
  transcript,
  gtf,
  association,
  hapmap,
  hapmap_ld = NULL,
  slide_length = -1,
  threadN = 1,
  up = NULL,
  down = NULL,
  threshold = NULL,
  ldstatistics = "rsquare",
  leadsnp = NULL,
  link2gene = NULL,
  triangleLD = TRUE,
  link2LD = NULL,
  leadsnpLD = TRUE,
  exon_colour = "gray",
  cds_colour = "black",
  utr_colour = "gray",
  intron_colour = "gray",
  colour02 = "gray",
  colour04 = "cyan",
  colour06 = "green",
  colour08 = "yellow",
  colour10 = "red",
  leadsnp_shape = 23,
  leadsnp_colour = "black",
  leadsnp_fill = "purple",
  leadsnp_size = 1.5,
  marker2highlight = NULL,
  marker2label = NULL,
  marker2label_angle = 60,
  marker2label_size = 1
)
}
\arguments{
\item{transcript}{the transcript of gene, required.}

\item{gtf}{the annotation file, required.}

\item{association}{the association table, required.}

\item{hapmap}{the genotype file for computing leadsnpLD in the format of hapmap. The file should be the same file used for coumputing association. required.}

\item{hapmap_ld}{the genotype file for computing trangleLD in the format of hapmap, not required. If hapmap_ld was not provided, hapmap would be used.}

\item{slide_length}{the sliding window length for computing LD, default -1.}

\item{threadN}{the number of (CPU) cores used for computing LD, default 1.}

\item{up}{the upper distance from the start position of gene}

\item{down}{the down distance from the end position of gene}

\item{threshold}{the significant level of the assocition, default NULL.}

\item{ldstatistics}{the statistics used for computing LD, default rsquare, and the optional is dprime.}

\item{leadsnp}{snp name provided by user}

\item{link2gene}{a dataframe speicify markers to be linked from GWAS to genic structure, default NULL. When link2gene is 'NULL', locis that passed the threshold will be linked. Please see help('marker2link').}

\item{triangleLD}{show LD in the format lile triangle, default TRUE.}

\item{link2LD}{a dataframe speicify markers to be linked from genic structure to LD matrix, default NULL. When link2gene is 'NULL', locis that passed the threshold will be linked. Please see help('marker2link').}

\item{leadsnpLD}{show LD of the locis when compared with the most significant loci, default TRUE.}

\item{exon_colour}{the colour of exon, default gray.}

\item{cds_colour}{the colour of cds, default black.}

\item{utr_colour}{the colour of utr, default gray.}

\item{intron_colour}{the colour of intron, default gray.}

\item{colour02}{the colour of LD statistics ranged between 0.0 and 0.2, default gray.}

\item{colour04}{the colour of LD statistics ranged between 0.2 and 0.4, default cyan.}

\item{colour06}{the colour of LD statistics ranged between 0.4 and 0.6, default green.}

\item{colour08}{the colour of LD statistics ranged between 0.6 and 0.8, default yellow.}

\item{colour10}{the colour of LD statistics ranged between 0.8 and 1.0, default red.}

\item{leadsnp_shape}{the shape of leadsnp, default 23. For others, please see help('points').}

\item{leadsnp_colour}{the shape of the point of leadsnp, default black. For others, please see help('points').}

\item{leadsnp_fill}{the filled colour of the point of leadsnp, default purple. For others, please see help('points').}

\item{leadsnp_size}{the size of of the point of leadsnp, default 1.5. For others, please see help('points').}

\item{marker2highlight}{a dataframe speicify markers to be showed by colour,shape,fill,size, default NULL. Please see help('marker2highlight').}

\item{marker2label}{a dataframe speicify markers to be labeled, default NULL. Please see help('marker2link')}

\item{marker2label_angle}{angel of labeled text, default 60.}

\item{marker2label_size}{size of labeled text, default 1.}
}
\value{
ggplot2 plot
}
\description{
This function plot the association with
linkage disequiblism and annotation at the level of a single gene.
}
\examples{
data(gtf)
data(zmvpp1_association)
data(zmvpp1_hapmap)
data(marker2highlight)
data(marker2link)
IntGenicPlot('GRMZM2G170927_T01',gtf,association=zmvpp1_association,hapmap=zmvpp1_hapmap,
hapmap_ld = zmvpp1_hapmap,threshold=8,up=500,down=600,leadsnpLD = FALSE,
marker2highlight=marker2highlight,link2gene=marker2link,link2LD=marker2link,
marker2label=marker2link,marker2label_angle=60,marker2label_size=2)
}
\author{
Hongwei Wang <\email{whweve@163.com}>
}
