% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntRegionalPlot.R
\name{IntRegionalPlot}
\alias{IntRegionalPlot}
\title{plot association with LD and annotation at a given region}
\usage{
IntRegionalPlot(
  chr,
  left,
  right,
  gtf,
  association,
  hapmap,
  hapmap_ld = NULL,
  slide_length = -1,
  threadN = 1,
  ldstatistics = "rsquare",
  leadsnp = NULL,
  threshold = NULL,
  link2gene = NULL,
  triangleLD = TRUE,
  link2LD = NULL,
  leadsnpLD = TRUE,
  label_gene_name = FALSE,
  colour02 = "gray",
  colour04 = "cyan",
  colour06 = "green",
  colour08 = "yellow",
  colour10 = "red",
  leadsnp_shape = 23,
  leadsnp_colour = "black",
  leadsnp_fill = "purple",
  leadsnp_size = 1.5,
  marker2highlight = NULL,
  marker2label = NULL,
  marker2label_angle = 60,
  marker2label_size = 1
)
}
\arguments{
\item{chr}{the chromosome, required.}

\item{left}{the left border of the region, required.}

\item{right}{the right border of the region, required.}

\item{gtf}{the annotation file, required.}

\item{association}{the association table, required.}

\item{hapmap}{the genotype file for computing leadsnpLD in the format of hapmap, required.}

\item{hapmap_ld}{the genotype file for computing trangleLD in the format of hapmap, not required. If hapmap_ld was not provided, hapmap would be used.}

\item{slide_length}{the sliding window length for computing LD, default -1.}

\item{threadN}{the number of (CPU) cores used for computing LD, default 1.}

\item{ldstatistics}{the statistics used for computing LD, default rsquare, and the optional is dprime.}

\item{leadsnp}{snp name provided by user}

\item{threshold}{the significant level of the assocition, default NULL.}

\item{link2gene}{a dataframe speicify markers to be linked from GWAS to genic structure, default NULL. When link2gene is 'NULL', locis that passed the threshold will be linked. Please see help('marker2link').}

\item{triangleLD}{show LD in the format lile triangle, default TRUE.}

\item{link2LD}{a dataframe speicify markers to be linked from genic structure to LD matrix, default NULL. When link2gene is 'NULL', locis that passed the threshold will be linked. Please see help('marker2link').}

\item{leadsnpLD}{show LD of the locis when compared with the most significant loci, default TRUE.}

\item{label_gene_name}{label the name of gene within the region, default FALSE.}

\item{colour02}{the colour of LD statistics ranged between 0.0 and 0.2, default gray.}

\item{colour04}{the colour of LD statistics ranged between 0.2 and 0.4, default cyan.}

\item{colour06}{the colour of LD statistics ranged between 0.4 and 0.6, default green.}

\item{colour08}{the colour of LD statistics ranged between 0.6 and 0.8, default yellow.}

\item{colour10}{the colour of LD statistics ranged between 0.8 and 1.0, default red.}

\item{leadsnp_shape}{the shape of leadsnp, default 23. For others, please see help('points').}

\item{leadsnp_colour}{the shape of the point of leadsnp, default black. For others, please see help('points').}

\item{leadsnp_fill}{the filled colour of the point of leadsnp, default purple. For others, please see help('points').}

\item{leadsnp_size}{the size of of the point of leadsnp, default 1.5. For others, please see help('points').}

\item{marker2highlight}{a dataframe speicify markers to be showed by colour,shape,fill,size, default NULL. Please see help('marker2highlight').}

\item{marker2label}{a dataframe speicify markers to be labeled, default NULL. Please see help('marker2link')}

\item{marker2label_angle}{angel of labeled text, default 60.}

\item{marker2label_size}{size of labeled text, default 1.}
}
\value{
ggplot2 plot
}
\description{
This function plot the association with
linkage disequiblism and annotation at the level of a region.
}
\examples{
data(gtf)
data(association)
data(hapmap_am368)
data(hapmap2)
IntRegionalPlot(chr=9,left=94178074-200000,right=94178074+200000,
gtf=gtf,association=association,hapmap=hapmap_am368,
hapmap_ld=hapmap_am368,threshold=5,leadsnp_size=2)
}
\author{
Hongwei Wang <\email{whweve@163.com}>
}
