% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker2highlight}
\alias{marker2highlight}
\title{Names of genotype name file.}
\format{:
\describe{
  \item{rs}{contains the SNP identifier.}
  \item{chrom}{contains the chromosome that the SNP was mapped.}
  \item{pos}{contains the respective position of this SNP on chromosome.}
  \item{shape}{contains the respective shape of this SNP on chromosome.}
  \item{colour}{contains the respective colour of this SNP on chromosome.}
  \item{fill}{contains the respective fill of this SNP on chromosome.}
  \item{size}{contains the respective size of this SNP on chromosome.}
}}
\usage{
marker2highlight
}
\description{
A dataset containing the name of genotype file. The variables are as follows:
}
\keyword{datasets}
