% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker2link}
\alias{marker2link}
\title{Names of genotype name file.}
\format{:
\describe{
  \item{rs}{contains the SNP identifier.}
  \item{chrom}{contains the chromosome that the SNP was mapped.}
  \item{pos}{contains the respective position of this SNP on chromosome.}
}}
\usage{
marker2link
}
\description{
A dataset containing the name of genotype file. The variables are as follows:
}
\keyword{datasets}
