% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zmvpp1_association}
\alias{zmvpp1_association}
\title{GWAS result.}
\format{An object of class \code{data.frame} with 124 rows and 4 columns.}
\usage{
zmvpp1_association
}
\description{
A dataset containing the association file from the resequenced ZmVpp1. The variables see association:
}
\keyword{datasets}
