% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zmvpp1_hapmap}
\alias{zmvpp1_hapmap}
\title{Genotype file.}
\format{An object of class \code{data.frame} with 124 rows and 152 columns.}
\usage{
zmvpp1_hapmap
}
\description{
A dataset containing the genotype file generated by resequence ZmVpp1. The variables see hapmap_am368
}
\keyword{datasets}
